# api-gateway/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Auction System API
  version: 1.0.0
  description: Production-grade auction platform — all endpoints
servers:
  - url: http://localhost:8080
    description: Local (API Gateway)
...
paths:
...
  /api/v1/auctions:
    get: {tags: [Auctions], summary: List auctions, parameters: [{name: state, in: query, schema: {type: string, enum: [active,scheduled,ended]}}, {name: page, in: query, schema: {type: integer}}, {name: page_size, in: query, schema: {type: integer}}], responses: {'200': {description: List}}}
    post: {tags: [Auctions], summary: Create auction (seller), security: [{BearerAuth: []}], requestBody: {required: true, content: {application/json: {schema: {type: object, required: [title,description,starting_price,start_time,end_time], properties: {title: {type: string}, description: {type: string}, starting_price: {type: number}, reserve_price: {type: number}, start_time: {type: string, format: date-time}, end_time: {type: string, format: date-time}}}}}}, responses: {'201': {description: Created}}}
  /api/v1/auctions/{id}:
    get: {tags: [Auctions], summary: Get auction, parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: {'200': {description: Auction}}}
    put: {tags: [Auctions], summary: Update auction, security: [{BearerAuth: []}], parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: {'200': {description: Updated}}}
    delete: {tags: [Auctions], summary: Delete auction, security: [{BearerAuth: []}], parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: {'200': {description: Deleted}}}
  /api/v1/auctions/{id}/publish:
    post: {tags: [Auctions], summary: Publish auction, security: [{BearerAuth: []}], parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: {'200': {description: Published}}}
  /api/v1/auctions/{id}/cancel:
    post: {tags: [Auctions], summary: Cancel auction, security: [{BearerAuth: []}], parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: {'200': {description: Cancelled}}}
  /api/v1/auctions/{id}/images:
    post: {tags: [Auction Images], summary: Add image, security: [{BearerAuth: []}], parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: {'201': {description: Added}}}
  /api/v1/auctions/{id}/images/{imageId}:
    delete: {tags: [Auction Images], summary: Delete image, security: [{BearerAuth: []}], parameters: [{name: id, in: path, required: true, schema: {type: string}}, {name: imageId, in: path, required: true, schema: {type: string}}], responses: {'200': {description: Deleted}}}
  /api/v1/seller/auctions:
    get: {tags: [Auctions], summary: My auctions (seller), security: [{BearerAuth: []}], responses: {'200': {description: Seller auctions}}}
...
  /api/v1/watchlist:
    get: {tags: [Watchlist], summary: My watchlist, security: [{BearerAuth: []}], responses: {'200': {description: Watchlist}}}
  /api/v1/watchlist/{auctionId}:
    post: {tags: [Watchlist], summary: Add to watchlist, security: [{BearerAuth: []}], parameters: [{name: auctionId, in: path, required: true, schema: {type: string}}], responses: {'201': {description: Added}}}
    delete: {tags: [Watchlist], summary: Remove from watchlist, security: [{BearerAuth: []}], parameters: [{name: auctionId, in: path, required: true, schema: {type: string}}], responses: {'200': {description: Removed}}}
...
  /api/v1/auctions/{id}/bids:
    get: {tags: [Bids], summary: Auction bids, parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: {'200': {description: Bids}}}
...
  /api/v1/wallet:
    get: {tags: [Wallet], summary: Get balance, security: [{BearerAuth: []}], responses: {'200': {description: Wallet}}}
  /api/v1/wallet/deposit:
    post: {tags: [Wallet], summary: Deposit, security: [{BearerAuth: []}], requestBody: {content: {application/json: {schema: {type: object, properties: {amount: {type: number}}}}}}, responses: {'200': {description: Deposited}}}
...
  /api/v1/wallet/history:
    get: {tags: [Wallet], summary: Transaction history, security: [{BearerAuth: []}], responses: {'200': {description: Transactions}}}
  /api/v1/payments/hold:
    post: {tags: [Wallet], summary: Hold balance for bid, security: [{BearerAuth: []}], requestBody: {content: {application/json: {schema: {type: object, properties: {auction_id: {type: string}, amount: {type: number}}}}}}, responses: {'200': {description: Held}}}
...
  /api/v1/notifications:
    get: {tags: [Notifications], summary: My notifications, security: [{BearerAuth: []}], responses: {'200': {description: Notifications}}}
  /api/v1/notifications/{id}/read:
    post: {tags: [Notifications], summary: Mark as read, security: [{BearerAuth: []}], parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: {'200': {description: Read}}}
  /api/v1/notifications/read-all:
    post: {tags: [Notifications], summary: Mark all read, security: [{BearerAuth: []}], responses: {'200': {description: All read}}}
...
  /api/v1/admin/auctions:
    get: {tags: [Admin], summary: All auctions (admin), security: [{BearerAuth: []}], responses: {'200': {description: Auctions}}}
  /api/v1/admin/auctions/{id}:
    delete: {tags: [Admin], summary: Force delete auction, security: [{BearerAuth: []}], parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: {'200': {description: Deleted}}}
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT